+++ src/main/java/um/tesoreria/core/hexagonal/dependencia/infrastructure/persistence/entity/DependenciaEntity.java
@@ -5,7 +5,7 @@ import lombok.*;
 import um.tesoreria.core.hexagonal.cuenta.infrastructure.persistence.entity.CuentaEntity;
 import um.tesoreria.core.hexagonal.geografica.infrastructure.persistence.entity.GeograficaEntity;
 import um.tesoreria.core.kotlin.model.Auditable;
-import um.tesoreria.core.kotlin.model.Facultad;
+import um.tesoreria.core.hexagonal.facultad.infrastructure.persistence.entity.FacultadEntity;
 import java.math.BigDecimal;
 
 @Getter
@@ -34,7 +34,7 @@ public class DependenciaEntity extends Auditable {
 
     @OneToOne(optional = true)
     @JoinColumn(name = "facultadId", insertable = false, updatable = false)
-    private Facultad facultad;
+    private FacultadEntity facultad;
 
     @OneToOne(optional = true)
     @JoinColumn(name = "geograficaId", insertable = false, updatable = false)