+++ src/main/java/um/tesoreria/core/model/view/ChequeraKey.java
@@ -20,6 +20,7 @@ import org.hibernate.annotations.Immutable;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import um.tesoreria.core.hexagonal.facultad.infrastructure.persistence.entity.FacultadEntity;
 import um.tesoreria.core.hexagonal.geografica.infrastructure.persistence.entity.GeograficaEntity;
 import um.tesoreria.core.hexagonal.persona.infrastructure.persistence.entity.PersonaEntity;
 import um.tesoreria.core.kotlin.model.*;
@@ -75,7 +76,7 @@ public class ChequeraKey extends Auditable implements Serializable {
 	
 	@OneToOne
 	@JoinColumn(name = "facultadId", insertable = false, updatable = false)
-	private Facultad facultad;
+	private FacultadEntity facultad;
 
 	@OneToOne
 	@JoinColumn(name = "tipoChequeraId", insertable = false, updatable = false)